drawGlow: compute depth-fade alpha in single precision (92.20->92.62, #139)
(f32) casts vs (double) drop the spurious fdiv and match target's
fctiwz/fsubs/fdivs single-prec sequence (one more fdivs/fsubs in target).

(cherry picked from commit b545742055c5bd3cb731d75fe8e578632eac6dfb)

+++ src/main/dll/dll_000A_expgfx.c
@@ -2477,8 +2477,8 @@ void drawGlow(u32 slotPoolBase, int poolIndex)
                 + ((f32*)viewMatrix)[11];
             if (viewProjW > lbl_803DB790)
             {
-                alpha = (int)((double)(s32)alpha * (double)((-viewProjW) - lbl_803DF414) /
-                    (double)((-lbl_803DB790) - lbl_803DF414));
+                alpha = (int)((f32)(s32)alpha * ((-viewProjW) - lbl_803DF414) /
+                    ((-lbl_803DB790) - lbl_803DF414));
             }
             *(volatile f32*)0xCC008000 = outX;
             *(volatile f32*)0xCC008000 = outY;
